refactor : path name 수정
